+++ src/platform.ts
@@ -35,8 +35,6 @@ export class MeaterPlatform implements DynamicPlatformPlugin {
   platformConfig!: MeaterPlatformConfig
   platformLogging!: options['logging']
   platformRefreshRate!: options['refreshRate']
-  platformPushRate!: options['pushRate']
-  platformUpdateRate!: options['updateRate']
   version: any
 
   constructor(
@@ -430,14 +428,6 @@ export class MeaterPlatform implements DynamicPlatformPlugin {
     this.platformRefreshRate = this.config.options?.refreshRate ? this.config.options.refreshRate : undefined
     const refreshRate = this.config.options?.refreshRate ? 'Using Platform Config refreshRate' : 'Platform Config refreshRate Not Set'
     await this.debugLog(`${refreshRate}: ${this.platformRefreshRate}`)
-    // UpdateRate
-    this.platformUpdateRate = this.config.options?.updateRate ? this.config.options.updateRate : undefined
-    const updateRate = this.config.options?.updateRate ? 'Using Platform Config updateRate' : 'Platform Config updateRate Not Set'
-    await this.debugLog(`${updateRate}: ${this.platformUpdateRate}`)
-    // PushRate
-    this.platformPushRate = this.config.options?.pushRate ? this.config.options.pushRate : undefined
-    const pushRate = this.config.options?.pushRate ? 'Using Platform Config pushRate' : 'Platform Config pushRate Not Set'
-    await this.debugLog(`${pushRate}: ${this.platformPushRate}`)
   }
 
   async getPlatformConfigSettings() {